+++ Projekt/Models/Portfolio.java
@@ -11,26 +11,26 @@ public class Portfolio {
     // TODO everything
 
     // flattened HashMap with composite key e.g. "Zurich:Nvidia"
-    private HashMap<String, Long> ownedStocks = new HashMap<String, Long>();
+    private HashMap<String, Integer> ownedStocks = new HashMap<String, Integer>();
 
-    private Long balance;
+    private Integer balance;
 
     final public String owner;
 
     public Portfolio(String owner) {
         this.owner = owner;
-        this.balance = 10000L;
+        this.balance = 10000;
     }
 
-    public Portfolio(String owner, Long balance) {
+    public Portfolio(String owner, Integer balance) {
         this.owner = owner;
         this.balance = balance;
     }
 
     public void purchase(StockMarket market, String stock, Integer amount) throws StockMarketException, UserInputException {
-        long amountPurchased = market.purchase(stock, amount);
-        long stockPrice = market.getPrice(stock);
-        long totalPrice = amountPurchased * stockPrice;
+        int amountPurchased = market.purchase(stock, amount);
+        int stockPrice = market.getPrice(stock);
+        int totalPrice = amountPurchased * stockPrice;
         balance -= totalPrice;
 
 
@@ -40,13 +40,13 @@ public class Portfolio {
             balance += totalPrice;
             throw new UserInputException("Insufficient funds to purchase this");
         } else {
-            Long currentStock = ownedStocks.get(market.getName()+":"+stock);
+            Integer currentStock = ownedStocks.get(market.getName()+":"+stock);
             currentStock = currentStock == null ? 0 : currentStock;
             ownedStocks.put(market.getName()+":"+stock, currentStock + amountPurchased);
         }
     }
 
-    public Long getBalance() { return this.balance; }
+    public Integer getBalance() { return this.balance; }
 
-    public HashMap<String, Long> getOwnedStocks() { return this.ownedStocks; }
+    public HashMap<String, Integer> getOwnedStocks() { return this.ownedStocks; }
 }
